Added function to construct designators that are used by `find-object'
In the function `find-object-with-id', an object designator is being constructed that can be used as a parameter to `find-object'. This object designator is specially tailored to include the gazebo model id that was associated with the model while spawning it. This way, `find-object' can find the gazebo model state and return the correct model pose. Atm, `find-object-with-id' calls `find-object' internally and returns the pose. This behavior might change in the future.