Added knowledge-backing to found objects and made function calls more clear
Split up the make-handled-object-designator function in one function creating the designator description from given parameters and one for creating the actual designator. This helps code-reuse as the designator description is needed elsewhere in the file as well. Also, added a function to fill up a designator with data from the knowledge base (`knowledge-backed-designator`) based on the object name. Added appropriate calls to `find-with-designator`.