Moved the gazebo object detection ros connection into cram-gazebo-utilities
As a circular dependency messed up the system, several functionalities were distributed to new packages. For once, the `get-model-pose' function (including the subscriber to the gazebo model state topic) was moved into the cram-gazebo-utilities package. Also, checking if whether a certain object name was spawned already was moved there. The gazebo-perception-process-module and the simple-knowledge packages were fitted accordingly as they depend on those functionalities. The circular dependency is therefore resolved and the cram-gazebo-utilities now holds all vital connection components with respect to ROS.